backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
python-dotenv
langgraph
pypdf2
